+++ WalletWasabi.Tests/UnitTests/Converters/OutPointAsTxoRefJsonConverterTests.cs
@@ -1,5 +1,3 @@
-using System;
-using System.ComponentModel.DataAnnotations;
 using NBitcoin;
 using Newtonsoft.Json;
 using WalletWasabi.JsonConverters;
@@ -15,7 +13,7 @@ namespace WalletWasabi.Tests.UnitTests
 			var outpoint = new OutPoint(new uint256("2716e680f47d74c1bc6f031da22331564dd4c6641d7216576aad1b846c85d492"), 765);
 			var txoref   = new TxoRef(new uint256("2716e680f47d74c1bc6f031da22331564dd4c6641d7216576aad1b846c85d492"), 765);
 
-			var oldConverters =new JsonConverter[]
+			var oldConverters = new JsonConverter[]
 			{
 				new Uint256JsonConverter(),
 				new OutPointJsonConverter(),
@@ -45,42 +43,5 @@ namespace WalletWasabi.Tests.UnitTests
 			var deserializedOutPoint2 = JsonConvert.DeserializeObject<TxoRef>(serializedOutPoint.ToLower(), newConverters);
 			Assert.Equal(deserializedOutPoint, deserializedOutPoint2);
 		}
-	}
-
-
-	[JsonObject(MemberSerialization.OptIn)]
-	class TxoRef
-	{
-		[Required]
-		[JsonProperty(Order = 1)]
-		[JsonConverter(typeof(Uint256JsonConverter))]
-		public uint256 TransactionId { get; }
-
-		[JsonProperty(Order = 2)]
-		public uint Index { get; }
-
-		[JsonConstructor]
-		public TxoRef(uint256 transactionId, uint index)
-		{
-			TransactionId = transactionId;
-			Index = index;
-		}
-
-		public override bool Equals(object obj)
-		{
-			if (ReferenceEquals(this, obj))
-			{
-				return true;
-			}
-
-			var other = obj as TxoRef;
-			if (this is null || other is null)
-			{
-				return false;
-			}
-			return (TransactionId, Index) == (other.TransactionId, other.Index);
-		}
-
-		public override int GetHashCode() => (TransactionId, Index).GetHashCode();
 	}
 }
